Refactor: First draft of new APIs
